move info button to trailing buttons widget
